chore: move configuration to config file
