+++ packages/config-resolver/src/regionConfig/index.ts
@@ -1,1 +1,2 @@
+export * from "./config";
 export * from "./resolveRegionConfig";
